Forcefully move the solution file to the root of the repository
This commit moves ./hmt/Halo-Mouse-Tool.sln to Halo-Mouse-Tool.sln, as
this ensures compatibility with the current structure of the repository.

The reason it was present in the hmt folder is because the previous
commit was cherry-picked from the SPV3 repository.

Additionally, this commit updates the project paths within the solution.

------------------------------------------------------------------------

NOTE: The solution file, for some reason, is declared to be ignored in
      the .gitignore file. Potential changes to the file might have to
      be forcefully added with the following command:

          git add -f Halo-Mouse-Tool.sln